tried to make a logon shell
but it sucked so I trashed it. since it's been a couple weeks, I figure I'd keep the patch in case it ends up being easier to use than starting from scratch. since it needed a place to stay we now have the .etc folder for non-asset, non-code files that are still of interest